backend/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
SQLAlchemy
psycopg2-binary
pandas
scikit-learn
nltk
openai
sentencepiece

# Pinning the specific versions that we know solve the build errors
numpy==1.26.4
torch==2.5.1
transformers